Fixes #25017 - Check which keys are passed into ext_search
Apparently when using an alias like 'user' (without the dot), it makes
scoped search pass *anything* starting with 'user.' into the external
search method. For this reason, we have to do another round of checking
if the key is actually allowed for searching.

The issue doesn't happen with aliases in the dot format (user.login).